Bugs fixed.

+++ cli/io.cpp
@@ -9,10 +9,10 @@ using namespace KaloMLib;
 void printMatrix(matrix &rhs)
 {
     printf("[\n");
-    for (int i = 0; i < rhs.ln; i++)
+    for (long i = 0; i < rhs.ln; i++)
     {
         printf("[");
-        for (int j = 0; j < rhs.col; j++)
+        for (long j = 0; j < rhs.col; j++)
             printf("%.4lf, ", rhs[i][j]);
         printf("],\n");
     }
@@ -22,12 +22,12 @@ void printMatrix(matrix &rhs)
 matrix inputMatrix()
 {
     printf("Please input the shape of the matrix : [n lns] [m cols]\n");
-    int n, m;
-    scanf("%d%d", &n, &m);
+    long n, m;
+    scanf("%ld%ld", &n, &m);
     printf("Please input the values of the matrix\n");
     matrix ret(make_pair(n, m));
-    for (int i = 0; i < n; i++)
-        for (int j = 0; j < m; j++)
+    for (long i = 0; i < n; i++)
+        for (long j = 0; j < m; j++)
             scanf("%lf", &ret[i][j]);
     return ret;
 }
